+++ back/src/main/java/ru/finam/backend/validation/ValidationService.java
@@ -16,14 +16,14 @@ public class ValidationService {
     }
 
     public void checkRequestDTOFieldsAreValid(FinanceInstrumentRequestDTO filter){
-        if( Double.compare(filter.getPriceFrom(), filter.getPriceUpTo()) > -1) {
+        if( Float.compare(filter.getPriceFrom(), filter.getPriceUpTo()) > -1) {
             throw new IllegalArgumentException("Некорректные данные для полей : priceFrom, priceUpTo");
         }
-        if(   Double.compare(filter.getCapitalizationFrom(), filter.getCapitalizationUpTo()) > -1){
+        if(   Float.compare(filter.getCapitalizationFrom(), filter.getCapitalizationUpTo()) > -1){
             throw new IllegalArgumentException("Некорректные данные для полей : " +
                     "capitalizationFrom, capitalizationUpTo");
         }
-        if( Double.compare(filter.getVolumeFrom(), filter.getVolumeUpTo()) > -1){
+        if( Float.compare(filter.getVolumeFrom(), filter.getVolumeUpTo()) > -1){
             throw new IllegalArgumentException("Некорректные данные для полей : volumeFrom, volumeUpTo");
         }
     }
